ENG-7943 Cloud: Stepfn/Lambda + many more (see details)
Fixes:
- Lambda waits for ec2 instance to enter become stable as
far as AWS SSM is concerned. This and several fixes in the
systemd scripts of xcalar, means the SF doesn't instantly
fail 70% of the time
- Events resources are always properly released now

New Features:
- Addition of `keep_cluster` bool in Events json objects, to keep
the cluster after a job
- Events job definition has access to the same env vars (eg,
${WORKBUCKET}) as the engine does
- xcalar_job.py has been extended to support these new feature,
including a dry-run mode where nothing is copied/deploy to your
events folder and just dumped to stdout
- New: Use ${LOGBUCKET} and ${LOGPREFIX} to store any custom logs

Improvements:
- Incorporate many learnings from 1000's of XcalarCloud
CF/lambda deployments
- Lots of cleanups in how the lambda CloudFormation
Templates (CFT) are built made for better separation of
concerns, which means simplified code
- Lambda layers were incorporated to help with DX
- Layers forced us to figure out consistent/repeatable
PIP packaging. These learnings have been incorporated to the
rest of the repo (XCE repo as well for XShell)
- You can debug/step through lambdas in the AWS Console editor
with full source code, because sizes reduction brought by layers
- Atmoic deploys of layers to enable multi-region support,
as per our offering requires
- CFT pipeline to go from templatized Jinja2 YAML CFT's to
to AWS SAM templates to JSON CFT
- Final CFT is linted/statically checked for errors before
deployment
- CFT manages all the lambda CloudWatchLogs (CWL) resources
properly now
- Image building pipeline
[Packer](https://jenkins.int.xcalar.com/job/Packer) takes
2/3 less time (from ~15min down ~6min)
- PIP package repeatablity is achieved via use of pip-compile
from the pip-tools package
- Integrate awscliv2 into XcalarCloud base images. This required
a bunch of related changes as the new v2 cli was created to
deal with Py2 EOL, which we still need around for CF
- Ec2: Any logged in user now has all the same env vars as XCE
does (WORKBUCKET, AWS_DEFAULT_REGION, etc)
- Ec2: xcalar-python36 package enabled "Persistent Pip" mode, where
pip modules are searched for in /mnt/xcalar/pysite, a major DX
improvement
- Naming improvements: Deployment urls, instance resources, and
env vars have all been looked over and made consistent where it
made sense. Add more details where it was missing (URL with build-
number of the build and the installer in it, for example)

+++ jenkins/Packer.sh
@@ -11,10 +11,14 @@ MANIFEST=$OUTDIR/packer-manifest.json
 export AWS_DEFAULT_REGION=${AWS_DEFAULT_REGION:-us-west-2}
 export PROJECT=${PROJECT:-xdp-awsmp}
 
-git clean -fxd aws packer azure
+if [ -n "$EXECUTOR_NUMBER" ]; then
+    git clean -fxd aws packer azure
+fi
 
 . infra-sh-lib
 . aws-sh-lib
+
+export VAULT_TOKEN=$($XLRINFRADIR/bin/vault-auth-puppet-cert.sh --print-token)
 
 if [ -n "$JENKINS_URL" ]; then
     if test -e .venv/bin/python2; then
@@ -34,14 +38,7 @@ source .venv/bin/activate || die "Failed to activate venv"
 
 test -d "$OUTDIR" || mkdir -p "$OUTDIR"
 
-do_packer() {
-    case "$BUILDER" in
-        amazon-*) CLOUD=aws; CLOUD_STORE=s3;;
-        arm-*|azure-*) CLOUD=azure; CLOUD_STORE=az;;
-        google*) CLOUD=google; CLOUD_STORE=gs;;
-        qemu*) CLOUD=qemu; CLOUD_STORE=;;
-    esac
-
+resolve_installer() {
     if [ -z "$INSTALLER_URL" ]; then
         if [ -d "$INSTALLER" ]; then
             echo "INSTALLER=$INSTALLER is a directory. Looking for an installer."
@@ -56,6 +53,25 @@ do_packer() {
             die "Failed to upload $INSTALLER to $CLOUD_STORE"
         fi
     fi
+    set -a
+    if [ -n "$INSTALLER" ]; then
+        eval "$(installer-version.sh --format=sh --image_build_number=$BUILD_NUMBER "$INSTALLER")"
+    elif [ -n "$INSTALLER_URL" ]; then
+        eval "$(installer-version.sh --format=sh --image_build_number=$BUILD_NUMBER "${INSTALLER_URL%%\?*}")"
+    fi
+    set +a
+}
+
+do_packer() {
+    case "$BUILDER" in
+        amazon-*) CLOUD=aws; CLOUD_STORE=s3;;
+        arm-*|azure-*) CLOUD=azure; CLOUD_STORE=az;;
+        google*) CLOUD=google; CLOUD_STORE=gs;;
+        qemu*) CLOUD=qemu; CLOUD_STORE=;;
+    esac
+
+    resolve_installer
+
     if [ -z "$LICENSE" ]; then
         if [ -z "$LICENSE_FILE" ]; then
             if build_is_rc "$INSTALLER"; then
@@ -75,10 +91,6 @@ do_packer() {
     cd $XLRINFRADIR/packer/aws
     #export INSTALLER_URL=$(installer-url.sh -d s3 $INSTALLER)
     #cfn-flip < $PACKERCONFIG > packer.json
-
-    INSTALLER_VERSION_BUILD=($(version_build_from_filename "$(filename_from_url "$INSTALLER_URL")"))
-    VERSION=${INSTALLER_VERSION_BUILD[0]}
-    RC=${INSTALLER_VERSION_BUILD[2]}
 
     unset INSTALLER
     export INSTALLER_URL
@@ -120,11 +132,17 @@ do_upload_template() {
     cd $XLRINFRADIR/aws/cfn
     local builder osid
     packer_manifest_all $MANIFEST > $OUTDIR/amis.yaml
+    if [ -z "$INSTALLER_TAG" ]; then
+        set -a
+        eval $(installer-version.sh --format=sh "$INSTALLER")
+        set +a
+    fi
     for builder in ${BUILDER//,/ }; do
         osid="${builder##*-}"
-        #if image_id="$(packer_ami_from_manifest ${builder} $MANIFEST)"; then
-        dc2 upload --project ${PROJECT} --manifest $MANIFEST --version ${VERSION} --release ${BUILD_NUMBER}${RELEASE:+.$RELEASE}-${osid} --url-file $OUTDIR/template-${builder}.url
-        #fi
+        dc2 upload --project ${PROJECT} --manifest $MANIFEST \
+            $(installer-version.sh --format=clieq "$INSTALLER") \
+            ${RELEASE:+--release ${RELEASE}} \
+            --url-file $OUTDIR/template-${builder}.url
     done
     )
 }
